Create header component
Created the header component for non-home page pages. Added this to
App.js to be rendered whenever the page is not the home page, which
should mean the header renders much less often then if it was
included in every view directly. Also added a placeholder for user
profile page to test the header and test redirect on Auth0
authorization.