Created protected route
Created a protected route Route component that will redirect
auth0 authenticated users to the Profile page. Also created the
logic to redirect a user without a profile to the pages to create
a profile -- will need to implement this as API endpoints become
available to perform and test this check.